Add Slack information to SlackOrganizationInfo
